% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_deg.R
\name{add_contact_labels}
\alias{add_contact_labels}
\title{Add interaction labels to spatial data}
\usage{
add_contact_labels(spatial, labels)
}
\arguments{
\item{spatial}{List of data frames containing spatial coordinates}

\item{labels}{List of neighbor information for each cell}
}
\value{
List of data frames with 'interactions' column added
}
\description{
This internal function annotates spatial coordinates with interaction information
by collapsing neighbor labels into concatenated strings.
}
\keyword{internal}
