% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_deg.R
\name{add_interactions}
\alias{add_interactions}
\title{Add interaction information to spatial data}
\usage{
add_interactions(
  spatial,
  cell_label = "cell_labels",
  collapse = FALSE,
  nn_method = "knn",
  k = 6
)
}
\arguments{
\item{spatial}{A list of data frames containing spatial coordinates}

\item{cell_label}{Character string specifying the column name for cell labels}

\item{collapse}{Logical indicating whether to collapse neighbor labels}

\item{nn_method}{Character string specifying the method for finding neighbors ("knn" or "voronoi")}

\item{k}{Integer specifying the number of nearest neighbors to consider (for knn method)}
}
\value{
A list of data frames with interaction information added
}
\description{
This function identifies cell-cell interactions in spatial data by finding
neighboring cells and annotates the spatial coordinates with interaction labels.
}
