% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rod_map.R
\name{dist_line}
\alias{dist_line}
\title{Calculate perpendicular distance from points to a line segment}
\usage{
dist_line(x_0, y_0, x_1, y_1, x, y)
}
\arguments{
\item{x_0}{Numeric value for x-coordinate of line start point}

\item{y_0}{Numeric value for y-coordinate of line start point}

\item{x_1}{Numeric value for x-coordinate of line end point}

\item{y_1}{Numeric value for y-coordinate of line end point}

\item{x}{Numeric vector of x-coordinates for points to measure}

\item{y}{Numeric vector of y-coordinates for points to measure}
}
\value{
Numeric vector of perpendicular distances from each point to the line segment
}
\description{
This internal function computes the perpendicular distance from multiple points
to a line segment defined by two endpoints.
}
\keyword{internal}
