% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_simulation}
\alias{export_simulation}
\title{Export simulated spatial data to CSV files}
\usage{
export_simulation(spatial, cells, out_dir = getwd(), file_tag = NULL)
}
\arguments{
\item{spatial}{A list of data frames containing spatial coordinates}

\item{cells}{A list of matrices containing gene count data}

\item{out_dir}{Character string specifying the output directory path}

\item{file_tag}{Character string to prefix output filenames}
}
\description{
This function exports the simulated spatial coordinates and gene count matrices
to CSV files for each sample in the simulation.
}
\details{
This function exports the simulated spatial coordinates and gene count matrices
to CSV files for each sample in the simulation.
}
\value{
NULL (files are written to disk)
}
