% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_cell_labels}
\alias{generate_cell_labels}
\title{Generate cell type labels for spatial territories}
\usage{
generate_cell_labels(spatial, cell_composition, randomize_cells = FALSE)
}
\arguments{
\item{spatial}{List of data frames containing spatial coordinates and territory assignments}

\item{cell_composition}{Integer specifying the number of cell types per territory}

\item{randomize_cells}{Logical indicating whether to randomize the number of cell types per territory}
}
\value{
Data frame with 'cell_labels' column added containing cell type assignments
}
\description{
This internal function assigns cell type labels to cells within spatial territories.
It can create multiple cell types per territory and optionally randomize cell type assignment.
}
\keyword{internal}
