% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_deg.R
\name{get_genes}
\alias{get_genes}
\title{Select genes for contact-dependent expression}
\usage{
get_genes(counts, interactions, n_genes)
}
\arguments{
\item{counts}{List of gene expression matrices}

\item{interactions}{Character vector of cell-cell interaction pairs}

\item{n_genes}{Integer specifying the number of genes to assign per interaction}
}
\value{
List where each element contains the selected genes for the corresponding interaction
}
\description{
This internal function randomly selects genes to be differentially expressed
for each cell-cell interaction pair.
}
\keyword{internal}
