% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_deg.R
\name{get_interactions}
\alias{get_interactions}
\title{Extract cell-cell interaction pairs from neighborhood data}
\usage{
get_interactions(spatial, nn_cells, cell_label = "cell_labels")
}
\arguments{
\item{spatial}{List of data frames containing spatial coordinates}

\item{nn_cells}{List of neighborhood information for each cell}

\item{cell_label}{Character string specifying the column name for cell labels}
}
\value{
Character vector of unique interaction pairs in format "celltype1@celltype2"
}
\description{
This internal function processes neighborhood information to identify all
unique cell-cell interaction pairs across the spatial data.
}
\keyword{internal}
