% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retrieve_counts}
\alias{retrieve_counts}
\title{Retrieve gene expression counts for spatial samples}
\usage{
retrieve_counts(spatial, sim)
}
\arguments{
\item{spatial}{List of data frames containing spatial coordinates with cell barcodes}

\item{sim}{SingleCellExperiment object containing simulated gene expression data}
}
\value{
List of matrices containing gene expression counts, one per spatial sample
}
\description{
This internal function extracts gene expression counts from a SingleCellExperiment
object and organizes them by spatial sample, matching barcodes to spatial coordinates.
}
\keyword{internal}
