% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chaos_maps.R
\name{tinkerbell}
\alias{tinkerbell}
\title{Generate tinkerbell strange attractor trajectory}
\usage{
tinkerbell(
  time = 8000,
  a = 0.9,
  b = -0.6013,
  c = 2,
  d = 0.5,
  x_0 = -0.72,
  y_0 = -0.64
)
}
\arguments{
\item{time}{Integer specifying the number of iterations to perform}

\item{a}{Numeric parameter for the tinkerbell map equation}

\item{b}{Numeric parameter for the tinkerbell map equation}

\item{c}{Numeric parameter for the tinkerbell map equation}

\item{d}{Numeric parameter for the tinkerbell map equation}

\item{x_0}{Numeric initial x-coordinate}

\item{y_0}{Numeric initial y-coordinate}
}
\value{
Data frame with 'x' and 'y' columns containing the trajectory coordinates
}
\description{
This internal function simulates the tinkerbell strange attractor by iteratively
applying the tinkerbell map equations to generate a chaotic trajectory.
}
\keyword{internal}
