% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chaos_maps.R
\name{chaos_map}
\alias{chaos_map}
\title{Create chaotic territoies}
\usage{
chaos_map(n_cells = 6000, max_expanse = 0.1, layers = 0, chaos = "tinkerbell")
}
\arguments{
\item{n_cells}{int - number of cells to create}

\item{max_expanse}{numeric -  max proportion of total width to use for
territory expansion.}

\item{layers}{int - number of layers in each rod territory}

\item{chaos}{character - string specifying which chaos map to use}
}
\value{
coordinate data frame with barcodes, x, y, and Territories
}
\description{
Create chaotic territoies
}
\details{
Create a a territory using chaos mapping means that there
be only one territory but it will have a chaotic shape.
}
