% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_map.R
\name{circular_map}
\alias{circular_map}
\title{Create circular  territoies}
\usage{
circular_map(
  n_cells = 6000,
  n_territories = 5,
  expanse = c(0.01, 0.5),
  layers = 0
)
}
\arguments{
\item{n_cells}{int - number of cells to create}

\item{n_territories}{int -  max number of territories to generate}

\item{expanse}{numeric -  max proportion of total width to use as 
circle radius}

\item{layers}{int - number of layers in each rod territory}
}
\value{
coordinate data frame with barcodes, x, y, and Territories
}
\description{
Create circular  territoies
}
