% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_simulated_data.R
\name{generate_sim_data}
\alias{generate_sim_data}
\title{Function to generate data for simuation analysis.}
\usage{
generate_sim_data(
  output,
  seed = 1453,
  plot = FALSE,
  run_mem = FALSE,
  simple = FALSE
)
}
\arguments{
\item{output}{path to directory where files will be added}

\item{seed}{seed for consitent generation}

\item{plot}{logical - export plots or not}
}
\description{
Function to generate data for simuation analysis.
}
\details{
The only purpose of this function is to output the data used
in the the manuscript. All parameters are set internally to force
to use these paremeters and these parameters only
}
