% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_spatial}
\alias{simulate_spatial}
\title{Create spatial territories}
\usage{
simulate_spatial(
  n_cells = 6000,
  n_territories = 5,
  n_samples = 10,
  pattern = "circle",
  layers = 0,
  max_expanse = 0.3,
  max_width = 0.3,
  max_length = 0.5,
  out_dir = NULL,
  border = TRUE
)
}
\arguments{
\item{n_cells}{int - number of cells to include in siluated data}

\item{n_territories}{int - number of territories to create}

\item{n_samples}{int - number of samples to generate}

\item{pattern}{character - string defining the type of territories to 
generate ("circle", "rod", "chaos")}

\item{layers}{int - number of layers in a territory}

\item{max_expanse}{numeric - expansion proportion for circular and chaotic
territories as a proportion of max size (1000 "pixels")}

\item{max_width}{numeric - max width of rod territories as a proportion of
max size (1000 "pixels")}

\item{max_length}{numeric - max length of rod territories as a proportion of
max size (1000 "pixels")}

\item{out_dir}{character - path to output directory. If null, will return list
of samples}

\item{border}{logical - should border cells have specific gene expression?}
}
\description{
Create spatial territories
}
