% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_spatial}
\alias{simulate_spatial}
\title{Create spatial territories}
\usage{
simulate_spatial(
  n_cells = 6000,
  n_territories = 5,
  n_samples = 10,
  pattern = "circle",
  expanse = c(0.01, 0.5),
  width_range = c(0, 0.3),
  length_range = c(0.1, 0.5),
  border = TRUE,
  layers = 0
)
}
\arguments{
\item{n_cells}{int - number of cells to include in siluated data}

\item{n_territories}{int - number of territories to create}

\item{n_samples}{int - number of samples to generate}

\item{pattern}{character - string defining the type of territories to 
generate ("circle", "rod", "chaos")}

\item{border}{logical - should border cells have specific gene expression?}

\item{layers}{int - number of layers in a territory}

\item{max_expanse}{numeric - expansion proportion for circular and chaotic
territories as a proportion of max size (1000 "pixels")}

\item{max_width}{numeric - max width of rod territories as a proportion of
max size (1000 "pixels")}

\item{max_length}{numeric - max length of rod territories as a proportion of
max size (1000 "pixels")}
}
\description{
Create spatial territories
}
